generate_X_df_multistates

This function generates a multistates CFD.

generate_X_df_multistates(
  nind = 100,
  N_states = 3,
  start = 0,
  end = 100,
  lambdas,
  transition_df,
  seed = 123
)

Arguments

nind

a int of the number of the individuals, default 100

N_states

a int of the number of states wanted, default 3

start

a value of starting time, default 0

end

a value of ending time, default 100

lambdas

a vector of N_states lambda values from lambda_determination(N_states)

transition_df

a dataframe with the transition matrix from tranfert_probabilities(N_states)

seed

a integer, random seed

Value

a dataframe of a multistates Categorical Funcitonal Data

Author

Francois Bassac

Examples

N_states = 4
lambdas = lambda_determination(N_states)
transition_df = transfert_probabilities(N_states)

df = generate_X_df_multistates(nind = 100, N_states, start=0, end=100,
lambdas,  transition_df)