naivePLS

This function performs the naive PLS method for Categorical functional data, Scalar functional data and multivariate data.

Usage

naivePLS(
  df_list,
  Y,
  regul_time_obj = NULL,
  curve_type_obj = NULL,
  id_col_obj = "id",
  time_col_obj = "time",
  print_steps = FALSE,
  plot_rmsep = TRUE,
  print_nbComp = TRUE,
  plot_reg_curves = FALSE,
  validation = "LOO",
  jackknife = TRUE
)

Arguments

df_list

a list of dataframe (id, time, value_or_state)

Y

a numeric vector for the scalar response

regul_time_obj

a list of time regularization values

curve_type_obj

a list of the curve types 'cat' or 'num'

id_col_obj

a list of character of the names of the id columns

time_col_obj

a list of character of the names of the time columns

print_steps

a boolean to print the different steps, default FALSE

plot_rmsep

a boolean to plot the RMSEP, default TRUE

print_nbComp

a boolean to print the optimal number or components, default TRUE

plot_reg_curves

a boolean to plot the regression curves, default FALSE

validation

a character, pls::plsr input, default 'LOO'

jackknife

a boolean, pls::plsr input, default TRUE

Value

a list of ("plsr_model", "nbCP_opti", "curves_names", "opti_reg_coef", "reg_obj")

Author

Francois Bassac