smoothPLS_multi_predict

This function use the list of regression functions to make a prediction \(\hat{Y} = \delta_0 + \int_0^T \delta(t) X(t) dt\)

smoothPLS_predict(
  df_predict_list,
  delta_list,
  curve_type_obj = NULL,
  id_col_obj = "id",
  time_col_obj = "time",
  regul_time_obj = NULL,
  int_mode = 1,
  nb_pt = 10,
  subdivisions = 100
)

Arguments

df_predict_list

a list of dataframe (id, time, value_or_state)

delta_list

a list of regression object (intercept, delta_1_fd, delta_2_fd, etc)

curve_type_obj

a list of characters of the curve types 'cat' or 'num'

id_col_obj

a list of character of the name of the id column, default 'id'

time_col_obj

a list of character of the name of the time column, default 'time'

regul_time_obj

a list of the time regularization values

int_mode

a integer for the integration mode, 1 for integrate, 2 for pracma::trapz

nb_pt

a integrer, number of intermediate points for pracma::trapz, default 10

subdivisions

a integer, number of subdivision in integrate function, default 100

Value

a numeric vector of the prediction

Author

Francois Bassac