assemble_basis_metric

This function assemble the metrics of all the basis of the basis list. This function only assemble the needed basis, especially if length(curve_to_keep) != N_states

Usage

assemble_basis_metric(basis_list, curves_to_keep = NULL)

Arguments

basis_list

a list of basis fd object

curves_to_keep

a list of the states curves to keep

Value

a matrix of the metric to consider

Author

Francois Bassac

Examples

basis1 = fda::create.bspline.basis(c(0,100), nbasis=10, norder=4)
basis2 = fda::create.bspline.basis(c(0,100), nbasis=15, norder=1)
basis3 = fda::create.fourier.basis(c(0,100), nbasis=7)
assemble_basis_metric(list(basis1, basis2, basis3), list(1,2,4))
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,] 2.04082039 1.250037112 0.263628199 0.017007716 0.000000000 0.000000000
#>  [2,] 1.25003711 3.163101206 2.232152219 0.493134361 0.004257681 0.000000000
#>  [3,] 0.26362820 2.232152219 4.668499367 3.208653988 0.338772911 0.002830051
#>  [4,] 0.01700772 0.493134361 3.208653988 6.847815210 3.375872551 0.340115340
#>  [5,] 0.00000000 0.004257681 0.338772911 3.375872551 6.848236092 3.375768459
#>  [6,] 0.00000000 0.000000000 0.002830051 0.340115340 3.375768459 6.848236092
#>  [7,] 0.00000000 0.000000000 0.000000000 0.002835021 0.340115340 3.375872551
#>  [8,] 0.00000000 0.000000000 0.000000000 0.000000000 0.002830051 0.338772911
#>  [9,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.004257681
#> [10,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [11,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [12,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [13,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [14,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [15,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [16,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [17,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [18,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [19,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [20,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [21,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [22,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [23,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [24,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [25,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [26,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [27,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [28,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [29,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [30,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [31,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [32,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#>              [,7]        [,8]        [,9]      [,10]    [,11]    [,12]    [,13]
#>  [1,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#>  [2,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#>  [3,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#>  [4,] 0.002835021 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#>  [5,] 0.340115340 0.002830051 0.000000000 0.00000000 0.000000 0.000000 0.000000
#>  [6,] 3.375872551 0.338772911 0.004257681 0.00000000 0.000000 0.000000 0.000000
#>  [7,] 6.847815210 3.208653988 0.493134361 0.01700772 0.000000 0.000000 0.000000
#>  [8,] 3.208653988 4.668499367 2.232152219 0.26362820 0.000000 0.000000 0.000000
#>  [9,] 0.493134361 2.232152219 3.163101206 1.25003711 0.000000 0.000000 0.000000
#> [10,] 0.017007716 0.263628199 1.250037112 2.04082039 0.000000 0.000000 0.000000
#> [11,] 0.000000000 0.000000000 0.000000000 0.00000000 6.666667 0.000000 0.000000
#> [12,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 6.666667 0.000000
#> [13,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 6.666667
#> [14,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [15,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [16,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [17,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [18,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [19,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [20,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [21,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [22,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [23,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [24,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [25,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [26,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [27,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [28,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [29,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [30,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [31,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [32,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#>          [,14]    [,15]    [,16]    [,17]    [,18]    [,19]    [,20]    [,21]
#>  [1,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [2,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [3,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [4,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [5,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [6,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [7,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [8,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [9,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [10,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [11,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [12,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [13,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [14,] 6.666667 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [15,] 0.000000 6.666667 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [16,] 0.000000 0.000000 6.666667 0.000000 0.000000 0.000000 0.000000 0.000000
#> [17,] 0.000000 0.000000 0.000000 6.666667 0.000000 0.000000 0.000000 0.000000
#> [18,] 0.000000 0.000000 0.000000 0.000000 6.666667 0.000000 0.000000 0.000000
#> [19,] 0.000000 0.000000 0.000000 0.000000 0.000000 6.666667 0.000000 0.000000
#> [20,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 6.666667 0.000000
#> [21,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 6.666667
#> [22,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [23,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [24,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [25,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [26,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [27,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [28,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [29,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [30,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [31,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [32,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>          [,22]    [,23]    [,24]    [,25]         [,26]         [,27]
#>  [1,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#>  [2,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#>  [3,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#>  [4,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#>  [5,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#>  [6,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#>  [7,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#>  [8,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#>  [9,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [10,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [11,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [12,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [13,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [14,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [15,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [16,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [17,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [18,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [19,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [20,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [21,] 0.000000 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [22,] 6.666667 0.000000 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [23,] 0.000000 6.666667 0.000000 0.000000  0.000000e+00  0.000000e+00
#> [24,] 0.000000 0.000000 6.666667 0.000000  0.000000e+00  0.000000e+00
#> [25,] 0.000000 0.000000 0.000000 6.666667  0.000000e+00  0.000000e+00
#> [26,] 0.000000 0.000000 0.000000 0.000000  1.000000e+00  4.994074e-17
#> [27,] 0.000000 0.000000 0.000000 0.000000  4.994074e-17  1.000000e+00
#> [28,] 0.000000 0.000000 0.000000 0.000000  7.162808e-17  6.491078e-18
#> [29,] 0.000000 0.000000 0.000000 0.000000 -1.408015e-17  6.782430e-17
#> [30,] 0.000000 0.000000 0.000000 0.000000  3.469081e-17 -2.618688e-17
#> [31,] 0.000000 0.000000 0.000000 0.000000  3.203884e-17 -1.383269e-06
#> [32,] 0.000000 0.000000 0.000000 0.000000  4.599502e-17 -3.121781e-18
#>               [,28]         [,29]         [,30]         [,31]         [,32]
#>  [1,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [2,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [3,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [4,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [5,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [6,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [7,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [8,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>  [9,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [10,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [11,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [12,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [13,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [14,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [15,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [16,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [17,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [18,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [19,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [20,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [21,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [22,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [23,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [24,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [25,]  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> [26,]  7.162808e-17 -1.408015e-17  3.469081e-17  3.203884e-17  4.599502e-17
#> [27,]  6.491078e-18  6.782430e-17 -2.618688e-17 -1.383269e-06 -3.121781e-18
#> [28,]  1.000000e+00  4.258708e-17  7.627178e-17 -1.022784e-16  1.383269e-06
#> [29,]  4.258708e-17  9.999986e-01 -9.152030e-17  1.489999e-16 -4.874251e-17
#> [30,]  7.627178e-17 -9.152030e-17  1.000001e+00 -9.014602e-17 -5.022236e-17
#> [31,] -1.022784e-16  1.489999e-16 -9.014602e-17  1.000000e+00  2.468871e-16
#> [32,]  1.383269e-06 -4.874251e-17 -5.022236e-17  2.468871e-16  1.000000e+00
assemble_basis_metric(list(basis1, basis2, basis3), list(1,2))
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,] 2.04082039 1.250037112 0.263628199 0.017007716 0.000000000 0.000000000
#>  [2,] 1.25003711 3.163101206 2.232152219 0.493134361 0.004257681 0.000000000
#>  [3,] 0.26362820 2.232152219 4.668499367 3.208653988 0.338772911 0.002830051
#>  [4,] 0.01700772 0.493134361 3.208653988 6.847815210 3.375872551 0.340115340
#>  [5,] 0.00000000 0.004257681 0.338772911 3.375872551 6.848236092 3.375768459
#>  [6,] 0.00000000 0.000000000 0.002830051 0.340115340 3.375768459 6.848236092
#>  [7,] 0.00000000 0.000000000 0.000000000 0.002835021 0.340115340 3.375872551
#>  [8,] 0.00000000 0.000000000 0.000000000 0.000000000 0.002830051 0.338772911
#>  [9,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.004257681
#> [10,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [11,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [12,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [13,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [14,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [15,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [16,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [17,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [18,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [19,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [20,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [21,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [22,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [23,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [24,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [25,] 0.00000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#>              [,7]        [,8]        [,9]      [,10]    [,11]    [,12]    [,13]
#>  [1,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#>  [2,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#>  [3,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#>  [4,] 0.002835021 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#>  [5,] 0.340115340 0.002830051 0.000000000 0.00000000 0.000000 0.000000 0.000000
#>  [6,] 3.375872551 0.338772911 0.004257681 0.00000000 0.000000 0.000000 0.000000
#>  [7,] 6.847815210 3.208653988 0.493134361 0.01700772 0.000000 0.000000 0.000000
#>  [8,] 3.208653988 4.668499367 2.232152219 0.26362820 0.000000 0.000000 0.000000
#>  [9,] 0.493134361 2.232152219 3.163101206 1.25003711 0.000000 0.000000 0.000000
#> [10,] 0.017007716 0.263628199 1.250037112 2.04082039 0.000000 0.000000 0.000000
#> [11,] 0.000000000 0.000000000 0.000000000 0.00000000 6.666667 0.000000 0.000000
#> [12,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 6.666667 0.000000
#> [13,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 6.666667
#> [14,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [15,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [16,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [17,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [18,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [19,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [20,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [21,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [22,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [23,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [24,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#> [25,] 0.000000000 0.000000000 0.000000000 0.00000000 0.000000 0.000000 0.000000
#>          [,14]    [,15]    [,16]    [,17]    [,18]    [,19]    [,20]    [,21]
#>  [1,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [2,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [3,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [4,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [5,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [6,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [7,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [8,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>  [9,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [10,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [11,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [12,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [13,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [14,] 6.666667 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [15,] 0.000000 6.666667 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [16,] 0.000000 0.000000 6.666667 0.000000 0.000000 0.000000 0.000000 0.000000
#> [17,] 0.000000 0.000000 0.000000 6.666667 0.000000 0.000000 0.000000 0.000000
#> [18,] 0.000000 0.000000 0.000000 0.000000 6.666667 0.000000 0.000000 0.000000
#> [19,] 0.000000 0.000000 0.000000 0.000000 0.000000 6.666667 0.000000 0.000000
#> [20,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 6.666667 0.000000
#> [21,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 6.666667
#> [22,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [23,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [24,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [25,] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>          [,22]    [,23]    [,24]    [,25]
#>  [1,] 0.000000 0.000000 0.000000 0.000000
#>  [2,] 0.000000 0.000000 0.000000 0.000000
#>  [3,] 0.000000 0.000000 0.000000 0.000000
#>  [4,] 0.000000 0.000000 0.000000 0.000000
#>  [5,] 0.000000 0.000000 0.000000 0.000000
#>  [6,] 0.000000 0.000000 0.000000 0.000000
#>  [7,] 0.000000 0.000000 0.000000 0.000000
#>  [8,] 0.000000 0.000000 0.000000 0.000000
#>  [9,] 0.000000 0.000000 0.000000 0.000000
#> [10,] 0.000000 0.000000 0.000000 0.000000
#> [11,] 0.000000 0.000000 0.000000 0.000000
#> [12,] 0.000000 0.000000 0.000000 0.000000
#> [13,] 0.000000 0.000000 0.000000 0.000000
#> [14,] 0.000000 0.000000 0.000000 0.000000
#> [15,] 0.000000 0.000000 0.000000 0.000000
#> [16,] 0.000000 0.000000 0.000000 0.000000
#> [17,] 0.000000 0.000000 0.000000 0.000000
#> [18,] 0.000000 0.000000 0.000000 0.000000
#> [19,] 0.000000 0.000000 0.000000 0.000000
#> [20,] 0.000000 0.000000 0.000000 0.000000
#> [21,] 0.000000 0.000000 0.000000 0.000000
#> [22,] 6.666667 0.000000 0.000000 0.000000
#> [23,] 0.000000 6.666667 0.000000 0.000000
#> [24,] 0.000000 0.000000 6.666667 0.000000
#> [25,] 0.000000 0.000000 0.000000 6.666667