funcPLS

This function performs the Multivariate Functional PLS as a matrix problem.

Usage

funcPLS(
  df_list,
  Y,
  basis_obj,
  regul_time_obj,
  curve_type_obj = NULL,
  id_col_obj = "id",
  time_col_obj = "time",
  print_steps = FALSE,
  plot_rmsep = TRUE,
  print_nbComp = TRUE,
  plot_reg_curves = FALSE,
  jackknife = TRUE,
  validation = "LOO"
)

Arguments

df_list

a list of dataframes (id, time, value_or_state)

Y

a numeric vector of the response

basis_obj

a basis fd obj or a list of basis fd obj. If basis fd obj, the same basis is used for all the curves

regul_time_obj

a vector of time regularization values or a list of vectors

curve_type_obj

a character "cat" or 'num' or a list of those values

id_col_obj

a character of the id column for all the curves or a list of id column character

time_col_obj

a character of the time column for all the curves or a list of time column character

print_steps

a boolean to cat the current step

plot_rmsep

a boolean to plot the plsr RMSEP

print_nbComp

a boolean to cat the optimal number of components

plot_reg_curves

a boolean to directly plot the beta regression curves

jackknife

a plsr input, default = TRUE

validation

a plsr input, default = 'LOO'

Value

a list ("curve_names", "alphas", "metric", "root_metric", "trans_alphas", "mfpls_mfd", "nb_comp_pls_opt", "beta_0", "beta_pls_list")

Author

Francois Bassac